requirements.txt
numpy>=1.24.2
wandb>=0.13.10
matplotlib>=3.7.0
scipy>=1.10.0
pytorch-lightning>=2.0.3
lightning==2.0.1.post
shapely>=2.0.1
networkx>=3.0
Cartopy>=0.22.0
pyproj>=3.4.1
tueplots>=0.0.8
plotly>=5.15.0
tornado>=6.3.3
cython>=3
cfgrib
xarray
argparse-dataclass==2.0.0
tensorboard==2.13.0
typer==0.9.0
netCDF4==1.6.5
tensorboard-plugin-profile 
torch-tb-profiler==0.4.1